# backend/internal/model/notification.go
package model

import "time"

type Notification struct {
	ID        string    `json:"id"`
	UserID    string    `json:"user_id"`
	Type      string    `json:"type"`
	PetID      *string   `json:"pet_id"`
	SightingID *string   `json:"sighting_id"`
	Title      string    `json:"title"`
	Body      string    `json:"body"`
	IsRead    bool      `json:"is_read"`
	CreatedAt time.Time `json:"created_at"`
}